requirements.txt
torch
torchvision
diffusers
numpy
pillow
opencv-python